print Pos as float with 0 dig after ,
Former-commit-id: e5fec2247a149f9b0775174e29ac0ad866cdb442

+++ include/ThDy_DNAHybrid.Nana/TableResults.h
@@ -82,7 +82,7 @@ class TableRes  : public nana::form, public EditableForm
     {
         float val(index row,  index col)const override
         {
-            return table->at(row,col )._Pos;
+            return static_cast<float>( table->at(row,col)._Pos );
         }
         Pos(Table &t) :value {t} {n_dec=0;};
    };
